+++ tests/test_entry_unload_refusal_offline.py
@@ -21,6 +21,7 @@ import logging
 
 import pytest
 from custom_components.haventory import services as services_mod
+from custom_components.haventory import subscriptions as subs_mod
 from custom_components.haventory import ws as ws_mod
 from custom_components.haventory.const import DOMAIN
 from custom_components.haventory.exceptions import NotLoadedError
@@ -197,7 +198,7 @@ async def test_unload_drops_live_subscriptions() -> None:
     await unload_entry(hass, entry)
     conn.messages.clear()
 
-    ws_mod.broadcast_event(hass, topic="items", action="created", payload={"item": {"id": "x"}})
+    subs_mod.broadcast_event(hass, topic="items", action="created", payload={"item": {"id": "x"}})
 
     assert conn.messages == []
     assert hass.data[DOMAIN].get("subscriptions") in (None, {})
@@ -230,7 +231,7 @@ async def test_teardown_signal_outranks_the_event_budget() -> None:
     # Drain the global budget, then prove an ordinary broadcast is now dropped.
     assert limiter.allow_event_broadcast() is True
     conn.messages.clear()
-    ws_mod.broadcast_event(hass, topic="items", action="created", payload={"item": {"id": "x"}})
+    subs_mod.broadcast_event(hass, topic="items", action="created", payload={"item": {"id": "x"}})
     assert conn.messages == []
 
     await unload_entry(hass, entry)
